'add_ddpg_beta'
