+++ examples/Pendulum/ddpg_pendulum.yaml
@@ -1,10 +1,10 @@
 seed: 0
-hidden_size: 30
-actor_lr: 0.00003
-critic_lr: 0.00003
+hidden_size: 32
+actor_lr: 0.001
+critic_lr: 0.002
 gamma: 0.9
-tau: 0.005
-var: 5
+tau: 0.01
+var: 2
 episode_length: 200
 mini_batch_size: 64
 buffer_size: 10000
